entidad para estadisticas de equipo del partido

+++ server/proyecto-fin-de-grado/src/main/java/com/ikerveiga/app/entity/TeamStats.java
@@ -187,7 +187,7 @@ public class TeamStats {
         return this.powerPlayPercentage;
     }
 
-    public void setPowerPlayPercentage(int powerPlayPercentage) {
+    public void setPowerPlayPercentage(float powerPlayPercentage) {
         this.powerPlayPercentage = powerPlayPercentage;
     }
